CMS-672: Update advisory sorting
